syndicatorapi/requirements.txt
beautifulsoup4==4.12.3
Django==5.0.2
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-ninja==1.1.0
email-normalize==2.0.0
lxml==5.1.0
requests==2.31.0
